hv:replace unsigned long long with uint64_t
unsigned long long--> uint64_t
long long --> int64_t

+++ hypervisor/include/lib/sprintf.h
@@ -30,7 +30,7 @@ struct print_param {
 		/* The parsed format precision. */
 		int precision;
 		/* The bitmask for unsigned values. */
-		unsigned long long mask;
+		uint64_t mask;
 		/* A pointer to the preformated value. */
 		const char *value;
 		/* The number of characters in the preformated value buffer. */
